requirements.txt
langchain==0.2.10
pypdf==4.3.0
chromadb==0.5.4
IPython==8.26.0
sentence_transformers==3.0.1
python-dotenv==1.0.1
flask==3.0.3
google-generativeai==0.7.2
markdown2==2.5.0
pandas==1.5.3
pandasai==2.2.12
langchain_google_genai